requirements.txt
Django==4.2.1
django-ckeditor==6.5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
mysqlclient==2.1.1
Pillow==9.5.0
python-decouple==3.8
uuid==1.30
